server/aarogyam-ml-server/app/requirements.txt
fastapi~=0.112.4
pymongo~=4.8.0
python-dotenv~=1.0.1
pydantic~=2.8.2
pyjwt~=2.6.0
numpy~=1.26.4
transformers~=4.44.2
requests~=2.32.3
beautifulsoup4~=4.12.3